Improve file format detection
Fixes #3

Currently does not detect the file format correctly if the input VCF filename is something like `samplename.vcf_altered.vcf.gz` where ".vcf" occurs multiple times.

This change fixes that by using substring removal instead of substring replacement to trim the extension from the filename. This has the advantage of being able to specify that the shortest match of ".vcf*" be removed, so additional ".vcf"s in the filename won't be affected. Technically the change to the `FILEFORMAT=` line isn't necessary, but I changed it just to be consistent with using substring removal instead of replacement.